Factorize test code
